+++ src/GlobalVariables.lua
@@ -1,4 +1,4 @@
 setfenv(1, MasterTradeSkills)
 
 MtsAddonName = "MasterTradeSkills";
-MtsAddonVersion = "2.4.2"
+MtsAddonVersion = "2.4.3"